Refactor XPath value and tokenizer modules
